requirements.txt
loguru==0.7.0
Pillow==9.5.0
watchdog==3.0.0
